+++ frontend/src/store/state.js
@@ -2,6 +2,7 @@ import { reactive } from 'vue';
 import { detectLocale } from "@/i18n";
 
 export const state = reactive({
+  showOverflowMenu: false,
   sessionId: "",
   disableOnlyOfficeExt: "",
   isSafari: /^((?!chrome|android).)*safari/i.test(navigator.userAgent),